BikeCsvImporter: import bikes, first step (main code, shop_id)
